requirements.txt
# ============================================================
# requirements.txt — InSAR Tĩnh Túc Project
# ============================================================
# Cài đặt: pip install -r requirements.txt
#
# Core (bắt buộc — thường đã có sẵn)
pyyaml>=6.0
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pandas>=2.0
scikit-learn>=1.3

# Deep learning (cho Transformer full — tùy chọn)
# Nếu không có, pipeline tự dùng NumPy linear baseline
torch>=2.0
torchvision>=0.15

# Wavelet Transform Coherence (cho ICA+WTC analysis)
PyWavelets>=1.4

# GIS & raster (cho production với dữ liệu thực)
geopandas>=0.14
shapely>=2.0
rasterio>=1.3
pyproj>=3.6

# SAR data I/O (khi tích hợp GAMMA/SNAP output)
h5py>=3.9
netCDF4>=1.6

# Notebook & reporting
# jupyter>=1.0
# ipywidgets>=8.0

# Testing
# pytest>=7.4
# pytest-cov>=4.1
